fix(ads): caste le total du feed en int (cache Redis renvoie une string en cache-hit)
Le feed /api/v1/ads renvoyait un 500 (SERVER_ERROR) sur toute requete
apres la premiere: RedisStore stocke les valeurs numeriques non serialisees
et les rend en string au cache-hit, donc $total arrivait en string au DTO
AdFeedResult (int strict) -> TypeError. Cast (int) a la sortie du cache.

Le premier appel (cache-miss) passait car la closure renvoie un int; les
tests ne l'avaient pas vu car CACHE_STORE=array round-trip l'int exact.
Ajoute un test de regression qui simule le cache-hit string.

+++ app/Services/Ad/AdFeedService.php
@@ -156,8 +156,11 @@ final readonly class AdFeedService
             );
         }
 
-        /** @var int $total */
-        $total = Cache::remember(
+        // Cast to int: the Redis cache store returns cached numeric values as
+        // strings on a cache-hit (RedisStore keeps numerics unserialized), so
+        // without this the string reaches AdFeedResult's strict `int $total`
+        // and throws a TypeError on every request after the first.
+        $total = (int) Cache::remember(
             'ads:feed:total:'.($typeId ?? 'all'),
             600,
             function () use ($typeId): int {
